# actual_ui/Code-Generation-UI/lib/api-zod/src/generated/types/memoSection.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoSectionStatus } from './memoSectionStatus';

export interface MemoSection {
  name: string;
  status: MemoSectionStatus;
  content: string;
  regenCount: number;
  reviewerNotes: string;
}