# actual_ui/Code-Generation-UI/lib/api-zod/src/generated/types/memoSectionActionInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MemoSectionActionInput {
  requestId: string;
  sectionName: string;
}