# actual_ui/Code-Generation-UI/lib/api-zod/src/generated/types/memoState.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MemoSection } from './memoSection';
import type { MemoStateStatus } from './memoStateStatus';

export interface MemoState {
  requestId: string;
  status: MemoStateStatus;
  approvedCount: number;
  totalSections: number;
  sections: MemoSection[];
  rationales: string[];
}