# actual_ui/Code-Generation-UI/lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  documentType: string;
  score: number;
  blobName: string;
  content: string;
}